Fully working contact list, connected to API!
Integrated a Redux store into the application, piloting a mechanism
for API calls I don't totally like quite yet (I think I may try and
set up a custom version of what fetch-action-creator does)

Also gave in and refactored the code into the redux-style layout
for actions and reducers. I may change that in the future but for
now it was the easiest way to get setup.