Added the outline of a spec for connecting the resume builder to the API
Basically, if you can get all these tests to pass, you should be like
70% of the way to having this working.

I would start with the reducer tests (reducers/resume.spec.js), since
that will get you some working state. At that point, if you use the
actions that are already written, the communication with the API should
work and happen cleanly (this has not yet been tested! You'll have to
do that)

The basic idea behind this spec is that the redux state is much closer
to the API representation (which is more general than the current
representation). This makes it easy to make sure the frontend doesn't
get out of sync with the backend (which is, in the long term, the
toughest thing to ensure in general).

The state in the Redux store is then generally going to be connected
to the resume builder components via the 'connect' call (see other
.container files for examples). This relies on the 'mapStateToProps'
function, which I've outlined some tests for here. You'll need to
create new '.container' files.

Note that generally this is going to replace the existing ResumeContext
state. I think it should be fairly clear how that should work, but
please let me know if you have questions about it.

I suggest when you are working on the mapStateToProps functions, that
you use a selector approach that takes advantage of the tools here:
https://redux-starter-kit.js.org/api/createselector

Also note that I didn't write a spec for how you would tie this stuff
to the ResumeBuilder class itself, since that class is currently the
furthest from actually complete, and I didn't want to speculate too
much. However, you should fairly easily be able to get the information
you need there from the existing state.